feat: Post Entity에 BaseTimeEntity 설정하기 [#1]

+++ src/main/kotlin/com/kotlin/study/dongambackend/domain/post/dto/response/PostResponse.kt
@@ -1,8 +1,8 @@
 package com.kotlin.study.dongambackend.domain.post.dto.response
 
 data class PostResponse(
-    var id: UInt,
-    var userId: UInt,   // TODO: Nested Object로 작성자 정보 전달하기
+    var id: Long,
+    var userId: Long,   // TODO: Nested Object로 작성자 정보 전달하기
     var title: String,
     var content: String,
     var category: String,
